[MIG] crm_facebook_leads: Migrate to 14.0
The following is performed:
- Fiels for models and views are reorganized by model name
- Fields containing an `_id` suffix but that were actually char are renamed,
  e.g. `facebook_form_id` -> `id_facebook_form`
- Lints are fixed, e.g. superfluous encoding headers
- And a little bit of code cleanup
