+++ crm_facebook_leads/__manifest__.py
@@ -1,31 +1,22 @@
-# -*- coding: utf-8 -*-
 {
     'name': "CRM Facebook Lead Ads",
-
     'summary': """
         Sync Facebook Leads with Odoo CRM""",
-
-    'description': """
-    """,
-
     'author': "BADEP, Vauxoo",
     'website': "https://badep.ma, www.vauxoo.com",
-
-    # Categories can be used to filter modules in modules listing
-    # Check https://github.com/odoo/odoo/blob/master/odoo/addons/base/module/module_data.xml
-    # for the full list
     'category': 'Lead Automation',
-    'version': '13.0.1.0.0',
-
-    # any module necessary for this one to work correctly
-    'depends': ['crm'],
-    'images': ['static/src/img/banner.png'],
+    'version': '14.0.1.0.0',
+    'depends': [
+        'crm',
+    ],
     'license': 'AGPL-3',
-
-    # always loaded
     'data': [
         'data/ir_config_parameter_data.xml',
+        'data/ir_cron_data.xml',
         'security/ir.model.access.csv',
-        'views/crm_view.xml',
-    ]
+        'views/crm_facebook_form_views.xml',
+        'views/crm_facebook_page_views.xml',
+        'views/crm_lead_views.xml',
+    ],
+    'application': True,
 }
