+++ crm_facebook_leads/data/ir_config_parameter_data.xml
@@ -1,3 +1,4 @@
+<?xml version="1.0"?>
 <odoo noupdate="1">
 
     <!-- Facebook Graph API Url Parameter -->
