+++ crm_facebook_leads/i18n/es.po
@@ -4,7 +4,7 @@
 #
 msgid ""
 msgstr ""
-"Project-Id-Version: Odoo Server 13.0+e\n"
+"Project-Id-Version: Odoo Server 14.0+e\n"
 "Report-Msgid-Bugs-To: \n"
 "POT-Creation-Date: 2020-12-10 02:32+0000\n"
 "PO-Revision-Date: 2020-12-10 02:32+0000\n"
@@ -52,7 +52,7 @@ msgstr "Nombre a Mostrar"
 #. module: crm_facebook_leads
 #: model:ir.model.fields,field_description:crm_facebook_leads.field_crm_lead__facebook_ad_id
 #: model:ir.model.fields,field_description:crm_facebook_leads.field_social_account__facebook_ad_id
-#: model:ir.model.fields,field_description:crm_facebook_leads.field_utm_medium__facebook_ad_id
+#: model:ir.model.fields,field_description:crm_facebook_leads.field_utm_medium__id_facebook_ad
 msgid "Facebook Ad"
 msgstr "Anuncio de Facebook"
 
@@ -64,8 +64,8 @@ msgstr "Grupo de Anuncios de Facebook"
 
 #. module: crm_facebook_leads
 #: model:ir.model.fields,field_description:crm_facebook_leads.field_crm_lead__facebook_campaign_id
-#: model:ir.model.fields,field_description:crm_facebook_leads.field_marketing_campaign__facebook_campaign_id
-#: model:ir.model.fields,field_description:crm_facebook_leads.field_utm_campaign__facebook_campaign_id
+#: model:ir.model.fields,field_description:crm_facebook_leads.field_marketing_campaign__id_facebook_campaign
+#: model:ir.model.fields,field_description:crm_facebook_leads.field_utm_campaign__id_facebook_campaign
 msgid "Facebook Campaign"
 msgstr "Campaña de Facebook"
 
@@ -148,7 +148,7 @@ msgid "Form"
 msgstr "Formulario"
 
 #. module: crm_facebook_leads
-#: model:ir.model.fields,field_description:crm_facebook_leads.field_crm_facebook_form__facebook_form_id
+#: model:ir.model.fields,field_description:crm_facebook_leads.field_crm_facebook_form__id_facebook_form
 msgid "Form ID"
 msgstr "ID de Formulario"
