+++ crm_facebook_leads/models/__init__.py
@@ -1,3 +1,6 @@
-# -*- coding: utf-8 -*-
-
-from . import lead
+from . import crm_facebook_form
+from . import crm_facebook_page
+from . import crm_lead
+from . import utm_adset
+from . import utm_campaign
+from . import utm_medium
