+++ crm_facebook_leads/models/utm_campaign.py
@@ -1,0 +1,11 @@
+from odoo import fields, models
+
+
+class UtmCampaign(models.Model):
+    _inherit = 'utm.campaign'
+
+    id_facebook_campaign = fields.Char(string="Facebook Campaign")
+
+    _sql_constraints = [
+        ('facebook_campaign_unique', 'unique(id_facebook_campaign)', 'This Facebook Campaign already exists!')
+    ]
