+++ crm_facebook_leads/views/crm_facebook_page_views.xml
@@ -1,0 +1,56 @@
+<?xml version="1.0"?>
+<odoo>
+
+    <record id="crm_facebook_page_form" model="ir.ui.view">
+        <field name="name">crm.facebook.page.form</field>
+        <field name="model">crm.facebook.page</field>
+        <field name="arch" type="xml">
+            <form string="Facebook Page">
+                <button name="get_forms" type="object" string="Get Forms"/>
+                <group colspan="2">
+                    <field name="name"/>
+                    <field name="access_token" password="1" class="oe_edit_only"/>
+                </group>
+                <group colspan="2">
+                    <field name="form_ids">
+                        <tree editable="0" readonly="1">
+                            <field name="name"/>
+                            <field name="allow_to_sync"/>
+                            <field name="id_facebook_form"/>
+                            <field name="team_id"/>
+                            <field name="campaign_id"/>
+                            <field name="medium_id"/>
+                            <field name="source_id"/>
+                        </tree>
+                    </field>
+                </group>
+            </form>
+        </field>
+    </record>
+
+    <record id="crm_facebook_page_tree" model="ir.ui.view">
+        <field name="name">crm.facebook.page.tree</field>
+        <field name="model">crm.facebook.page</field>
+        <field name="arch" type="xml">
+            <tree>
+                <field name="name"/>
+            </tree>
+        </field>
+    </record>
+
+    <record id="action_crm_facebook_page" model="ir.actions.act_window">
+        <field name="name">Facebook Pages</field>
+        <field name="res_model">crm.facebook.page</field>
+        <field name="view_mode">tree,form</field>
+        <field name="view_id" ref="crm_facebook_page_tree"/>
+    </record>
+
+    <menuitem
+        id="menu_crm_facebook_page"
+        name="Facebook Pages"
+        parent="crm.menu_crm_config_lead"
+        sequence="100"
+        action="action_crm_facebook_page"
+        groups="sales_team.group_sale_manager"/>
+
+</odoo>
