+++ requirements.txt
@@ -1,0 +1,1 @@
+openupgradelib @ git+https://github.com/OCA/openupgradelib.git
